requirements.txt
## library module requirements
pyyaml
packaging
